PR #103 round 1: address CodeRabbit + Sonar review feedback
CodeRabbit (8 valid threads):
- ingestion.py:413 — INT-named unmapped types (MEDIUMINT/etc.) now narrow to INT, not the coarse DOUBLE fallback.
- yaml_storage.py / sqlite_storage.py — always persist when v4→v5 migration runs, even if the live-schema refinement makes no changes (text-only models, query-backed, all-DOUBLE columns). Without this fix the on-disk file re-migrates / re-introspects on every load.
- enrichment.py + enriched.py + generator.py — propagate ``qfield.type`` to ``EnrichedTransform`` (lag/lead/cumsum/time_shift/rank/...) and emit CAST in the window-layer + final-SELECT transform paths. DEV-1361 type-aware emission is now complete for the transform branch too.
- cli.py — drop the duplicate ``_add_storage_arg`` on the ``storage`` parent (subcommand owns ``--storage`` per existing CLI patterns).
- CLAUDE.md — replace legacy ``type=NUMBER`` example with ``type=DataType.DOUBLE``; add a ``slayer storage migrate-types`` bullet to the CLI section.
- test_mcp_inspect.py fixtures — INTEGER-backed PK columns and ``quantity`` declared as ``DataType.INT``; ``label`` as ``DataType.TEXT`` (matched the table schema, fixed an inconsistency).
- test_integration.py json_extract test — pinned CAST emission via a ``dry_run=True`` round-trip assertion (the runtime tuple comparison passed with or without CAST on this fixture).
- test_validate_models.py INT-vs-DOUBLE bucket test — actually exercises ``DataType.INT`` on the live side now.

CodeRabbit (3 invalid replies posted): outdated threads on DEV-1356 territory (server.py ingest exception handling, query_engine touched-model collection, schema_drift join-col resolution).

Sonar (10 issues, 1 refactor + 9 NOSONAR):
- cli.py S3776 — extracted ``_refine_one_model_for_cli`` from ``_run_storage_migrate_types`` to drop cognitive complexity below 15.
- cli.py:638 S7493 — sync I/O in async by design (CLAUDE.md, Async Architecture).
- test_storage_type_refinement.py S2068 — ``password='nope'`` is a test fixture targeting a closed port to assert hard-fail.
- test_storage_type_refinement.py S7493 (×6) — sync I/O in test fixtures.
- test_storage_type_refinement.py S112 — ``Exception(...)`` is the cause-of arg for the simulated SQLAlchemy connect error.
- test_storage_type_refinement.py S7503 (×2) — pytest-asyncio test bodies.

DEV-1361 follow-up: added ``EnrichedMeasure.column_type`` so the inner (pre-aggregation) expression of an aggregation gets CAST'd when the source ``Column.type`` is set on a non-bare ``sql`` (e.g. ``SUM(CAST(json_extract(payload, '$.score') AS REAL))``). This was missing before — column-level CAST only fired through the dimension path. ``_wrap_cast_for_type`` also skips bare ``exp.Column`` nodes (including qualified ``model.col`` after column-expansion) to avoid lossy casts on SQLite text-stored timestamps.

+++ slayer/engine/ingestion.py
@@ -410,7 +410,10 @@ def _get_columns_fallback(
             sa_type = sa_type or DataType.DOUBLE
             is_float = _parse_info_schema_is_float(data_type_str)
         elif sa_type is None and "INT" in base_type:
-            sa_type = DataType.DOUBLE
+            # DEV-1361: integer-shaped types should narrow to INT, not the
+            # coarse DOUBLE fallback (e.g. MEDIUMINT, TINYINT variants not
+            # otherwise mapped).
+            sa_type = DataType.INT
         elif sa_type is None and ("CHAR" in base_type or "TEXT" in base_type):
             sa_type = DataType.TEXT
         result.append({"name": col_name, "type": sa_type or DataType.TEXT, "is_float": is_float})
